# Implement performance benchmarks