--- a/.github/workflows/xcmetrics.yml
+++ b/.github/workflows/xcmetrics.yml
@@ -6,6 +6,9 @@ on:
     - cron: '0 3 * * *'
 
   pull_request:
+    branches: # FIXME: delete this
+      - '**'
+
     types:
       - opened
       - ready_for_review