--- a/Sources/StreamChatSwiftUI/Appearance.swift
+++ b/Sources/StreamChatSwiftUI/Appearance.swift
@@ -48,3 +48,5 @@ extension EnvironmentValues {
         }
     }
 }
+
+// I'M TESTING PERFORMACE BENCHMARKS