--- a/Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListConfig.swift
+++ b/Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListConfig.swift
@@ -28,7 +28,8 @@ public struct MessageListConfig {
         handleTabBarVisibility: Bool = true,
         messageListAlignment: MessageListAlignment = .standard,
         uniqueReactionsEnabled: Bool = false,
-        localLinkDetectionEnabled: Bool = true
+        localLinkDetectionEnabled: Bool = true,
+        isMessageEditedLabelEnabled: Bool = true
     ) {
         self.messageListType = messageListType
         self.typingIndicatorPlacement = typingIndicatorPlacement
@@ -50,6 +51,7 @@ public struct MessageListConfig {
         self.messageListAlignment = messageListAlignment
         self.uniqueReactionsEnabled = uniqueReactionsEnabled
         self.localLinkDetectionEnabled = localLinkDetectionEnabled
+        self.isMessageEditedLabelEnabled = isMessageEditedLabelEnabled
     }
 
     public let messageListType: MessageListType
@@ -72,6 +74,7 @@ public struct MessageListConfig {
     public let messageListAlignment: MessageListAlignment
     public let uniqueReactionsEnabled: Bool
     public let localLinkDetectionEnabled: Bool
+    public let isMessageEditedLabelEnabled: Bool
 }
 
 /// Contains information about the message paddings.