--- a/StreamChatSwiftUI.xcodeproj/project.pbxproj
+++ b/StreamChatSwiftUI.xcodeproj/project.pbxproj
@@ -9,6 +9,8 @@
 /* Begin PBXBuildFile section */
 		402C54482B6AAC0100672BFB /* StreamChatSwiftUI.framework in Frameworks */ = {isa = PBXBuildFile; fileRef = 8465FBB52746873A00AF091E /* StreamChatSwiftUI.framework */; };
 		402C54492B6AAC0100672BFB /* StreamChatSwiftUI.framework in Embed Frameworks */ = {isa = PBXBuildFile; fileRef = 8465FBB52746873A00AF091E /* StreamChatSwiftUI.framework */; settings = {ATTRIBUTES = (CodeSignOnCopy, RemoveHeadersOnCopy, ); }; };
+		8204344E2B973B7F00C5B94A /* ChannelListScrollTime.swift in Sources */ = {isa = PBXBuildFile; fileRef = 8204344D2B973B7F00C5B94A /* ChannelListScrollTime.swift */; };
+		820434502B973B8A00C5B94A /* MessageListScrollTime.swift in Sources */ = {isa = PBXBuildFile; fileRef = 8204344F2B973B8A00C5B94A /* MessageListScrollTime.swift */; };
 		8205B4142AD41CC700265B84 /* StreamSwiftTestHelpers in Frameworks */ = {isa = PBXBuildFile; productRef = 8205B4132AD41CC700265B84 /* StreamSwiftTestHelpers */; };
 		8205B4182AD4267200265B84 /* StreamSwiftTestHelpers in Frameworks */ = {isa = PBXBuildFile; productRef = 8205B4172AD4267200265B84 /* StreamSwiftTestHelpers */; };
 		820A61A029D6D78E002257FB /* QuotedReply_Tests.swift in Sources */ = {isa = PBXBuildFile; fileRef = 820A619F29D6D78E002257FB /* QuotedReply_Tests.swift */; };
@@ -546,6 +548,9 @@
 
 /* Begin PBXFileReference section */
 		4A65451E274BA170003C5FA8 /* README.md */ = {isa = PBXFileReference; lastKnownFileType = net.daringfireball.markdown; path = README.md; sourceTree = "<group>"; };
+		8204344B2B97396300C5B94A /* Performance.xctestplan */ = {isa = PBXFileReference; lastKnownFileType = text; path = Performance.xctestplan; sourceTree = "<group>"; };
+		8204344D2B973B7F00C5B94A /* ChannelListScrollTime.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = ChannelListScrollTime.swift; sourceTree = "<group>"; };
+		8204344F2B973B8A00C5B94A /* MessageListScrollTime.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = MessageListScrollTime.swift; sourceTree = "<group>"; };
 		820A619F29D6D78E002257FB /* QuotedReply_Tests.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = QuotedReply_Tests.swift; sourceTree = "<group>"; };
 		825AADF3283CCDB000237498 /* ThreadPage.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = ThreadPage.swift; sourceTree = "<group>"; };
 		829AB4D128578ACF002DC629 /* StreamTestCase+Tags.swift */ = {isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = "StreamTestCase+Tags.swift"; sourceTree = "<group>"; };
@@ -1071,6 +1076,15 @@
 /* End PBXFrameworksBuildPhase section */
 
 /* Begin PBXGroup section */
+		8204344C2B973B6D00C5B94A /* Performance */ = {
+			isa = PBXGroup;
+			children = (
+				8204344D2B973B7F00C5B94A /* ChannelListScrollTime.swift */,
+				8204344F2B973B8A00C5B94A /* MessageListScrollTime.swift */,
+			);
+			path = Performance;
+			sourceTree = "<group>";
+		};
 		82A1814528FD69E8005F9D43 /* Message Delivery Status */ = {
 			isa = PBXGroup;
 			children = (
@@ -1313,6 +1327,7 @@
 			isa = PBXGroup;
 			children = (
 				A3828EB0283F73EE00538258 /* StreamChatSwiftUITestsApp.xctestplan */,
+				8204344B2B97396300C5B94A /* Performance.xctestplan */,
 				A3600B36283E9EC900E1C930 /* StreamChatSwiftUITests.swift */,
 				A3600B33283E9EBA00E1C930 /* Extensions */,
 				8463D92428365E8F002B1894 /* Pages */,
@@ -2056,14 +2071,15 @@
 		A3600B30283E9E4100E1C930 /* Tests */ = {
 			isa = PBXGroup;
 			children = (
+				8204344C2B973B6D00C5B94A /* Performance */,
 				A3600B3F283F651200E1C930 /* Base TestCase */,
+				82A1814528FD69E8005F9D43 /* Message Delivery Status */,
 				82A1813D28FD68A3005F9D43 /* ChannelList_Tests.swift */,
 				A3600B31283E9E4700E1C930 /* MessageList_Tests.swift */,
 				829AB4D32858A532002DC629 /* Reactions_Tests.swift */,
 				82A1813B28F9BA53005F9D43 /* Attachments_Tests.swift */,
 				82A1814328FD69AE005F9D43 /* SlowMode_Tests.swift */,
 				829AB4D128578ACF002DC629 /* StreamTestCase+Tags.swift */,
-				82A1814528FD69E8005F9D43 /* Message Delivery Status */,
 				82A1813F28FD691B005F9D43 /* Ephemeral_Messages_Tests.swift */,
 				82A1814128FD694A005F9D43 /* PushNotification_Tests.swift */,
 				820A619F29D6D78E002257FB /* QuotedReply_Tests.swift */,
@@ -2355,6 +2371,8 @@
 				825AADF4283CCDB000237498 /* ThreadPage.swift in Sources */,
 				A3828EAD283F6CFE00538258 /* StartPage.swift in Sources */,
 				A3600B41283F652400E1C930 /* StreamTestCase.swift in Sources */,
+				8204344E2B973B7F00C5B94A /* ChannelListScrollTime.swift in Sources */,
+				820434502B973B8A00C5B94A /* MessageListScrollTime.swift in Sources */,
 				82A1814928FD6A0C005F9D43 /* MessageDeliveryStatus+ChannelList_Tests.swift in Sources */,
 				82A1814228FD694A005F9D43 /* PushNotification_Tests.swift in Sources */,
 				82A1814428FD69AE005F9D43 /* SlowMode_Tests.swift in Sources */,
@@ -3375,8 +3393,8 @@
 			isa = XCRemoteSwiftPackageReference;
 			repositoryURL = "https://github.com/GetStream/stream-chat-swift.git";
 			requirement = {
-				kind = upToNextMajorVersion;
-				minimumVersion = 4.49.0;
+				branch = develop;
+				kind = branch;
 			};
 		};
 		E3A1C01A282BAC66002D1E26 /* XCRemoteSwiftPackageReference "sentry-cocoa" */ = {