--- a/StreamChatSwiftUI.xcodeproj/xcshareddata/xcschemes/StreamChatSwiftUITestsApp.xcscheme
+++ b/StreamChatSwiftUI.xcodeproj/xcshareddata/xcschemes/StreamChatSwiftUITestsApp.xcscheme
@@ -24,13 +24,16 @@
    </BuildAction>
    <TestAction
       buildConfiguration = "Debug"
-      selectedDebuggerIdentifier = ""
-      selectedLauncherIdentifier = "Xcode.IDEFoundation.Launcher.PosixSpawn"
+      selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
+      selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <TestPlans>
          <TestPlanReference
             reference = "container:StreamChatSwiftUITestsAppTests/StreamChatSwiftUITestsApp.xctestplan"
             default = "YES">
+         </TestPlanReference>
+         <TestPlanReference
+            reference = "container:StreamChatSwiftUITestsAppTests/Performance.xctestplan">
          </TestPlanReference>
       </TestPlans>
       <Testables>
@@ -48,8 +51,8 @@
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
-      selectedDebuggerIdentifier = ""
-      selectedLauncherIdentifier = "Xcode.IDEFoundation.Launcher.PosixSpawn"
+      selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
+      selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
       ignoresPersistentStateOnLaunch = "NO"