--- a/StreamChatSwiftUITestsAppTests/StreamChatSwiftUITestsApp.xctestplan
+++ b/StreamChatSwiftUITestsAppTests/StreamChatSwiftUITestsApp.xctestplan
@@ -19,6 +19,10 @@
   },
   "testTargets" : [
     {
+      "skippedTests" : [
+        "ChannelListScrollTime",
+        "MessageListScrollTime"
+      ],
       "target" : {
         "containerPath" : "container:StreamChatSwiftUI.xcodeproj",
         "identifier" : "8400A350282E6BE30067D3A0",