### 🔗 Issue Links

Resolves https://github.com/GetStream/ios-issues-tracking/issues/659

### 🎯 Goal

Implement Performance Benchmarks

### 🔖 Notion
- https://www.notion.so/stream-wiki/XCMetrics-5ef80c5ae9d44772966ab99f6c6d094a

### 📝 Summary

- https://github.com/GetStream/stream-swift-performance-benchmarks stores the benchmarks
- XCMetrics are executed on every PR (including release) by `first PR's approval`, `manually on any branch` and `nightly by cron`
- XCMetrics won't be executed if there are no changes to the `Sources` folder
- Collected metrics: `:hitches_total_duration`, `:duration`, `:hitch_time_ratio`, `:frame_count`, `:frame_rate`, `:number_of_hitches`
- Metrics are collected from real devices on Firebase TestLab. Firebase offers us 5 runs per day (although, for some reason, sometimes it's even more than that), which should be enough most of the time (`1 cron, 4 PRs`). If this isn't enough, we can set up a paid account that offers free 30 minutes per day (`1run ≈ 4min`)
- Our custom XCMetrics threshold equals to Xcode default one - `10%`
- Performance tests are executed on `MessageList` (with mocked static data: messages/images/videos/pdfs/giphies) and `ChannelList` (without mocks)
- As far as I understood, Apple requires developers to run performance tests on the same machine + same iPhone ([read more](https://developer.squareup.com/blog/measureblock-how-does-performance-testing-work-in-ios/#qa)), which is impossible on CI. That's why there is no configured baseline in this PR, and the XCTests accordingly will always pass. Worries aside, because I implemented a custom calculation & assertion on the fastlane level.

### 🎨 Showcase

| `target` | `metric` | `branch` | `release/4.48.1` | `performance` | `status` |
| - | - | - | - | - | - |
| MessageList | Hitches total duration | 5.0 ms | 20 ms | 75.0% 🔼 | 🟢 |
|  | Duration | 2.6 s | 5 s | 48.0% 🔼 | 🟢 |
|  | Hitch time ratio | 2.0 ms per s | 10 ms per s | 80.0% 🔼 | 🟢 |
|  | Frame count | 199  | 300  | 33.7% 🔼 | 🟢 |
|  | Frame rate | 78.4 fps | 100 fps | 21.6% 🔼 | 🟢 |
|  | Number of hitches | 0  | 1  | 100.0% 🔼 | 🟢 |
| ChannelList | Hitches total duration | 10.0 ms | 20 ms | 50.0% 🔼 | 🟢 |
|  | Duration | 2.6 s | 5 s | 48.0% 🔼 | 🟢 |
|  | Hitch time ratio | 3.9 ms per s | 10 ms per s | 61.0% 🔼 | 🟢 |
|  | Frame rate | 74.7 fps | 100 fps | 25.3% 🔼 | 🟢 |
|  | Number of hitches | 0  | 1  | 100.0% 🔼 | 🟢 |

### 🧪 Manual Testing Notes

- `bundle exec fastlane xcmetrics`

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change follows zero ⚠️ policy (required)
- [ ] This change should be manually QAed
- [ ] Changelog is updated with client-facing changes
- [ ] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

### 🎁 Meme

![gif](https://media.giphy.com/media/v1.Y2lkPTc5MGI3NjExZ3RsazY3b3g3MnRzMWw0c2w2OGYxOHRkdGU3bGVsNTF1MTVwOTh3MCZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/2MoQlOxVx5TgY/giphy.gif)